Fixed bug in PDFMixingModel so that PDF mixing tables can be read in. Cleaned up some code to get rid of compiler warnings.

+++ Mixing/PDFMixingModel.h
@@ -117,7 +117,7 @@ public:
       inline int getTableDimension() const{
 	return d_tableDimension;
       }
-      inline int getStoicPt() const{ 
+      inline double getStoicPt() const{ 
 	return d_stoicValue; 
       }
       inline std::string getMixTableType() const{
